:bug: Fix double-mounted Dashboard view
The DashboardView component was actually being mounted twice, resulting
in two calls to the database to create a user, one of which would fail
and cause the app to hang. The issue was that the Authenticate HOC that DashboardView was being called with wasn't extending React.Component,
but rather extending the Component passed into it. Specifying
React.Component fixes the issue. (hat tip to @ajb85 for the catch)