:bug: debuged and got stripe frontend code working again
created a stripe acccount for this app. created products for production and for testing in stripe and different payment plans and subscriptions for those products. Replaced the plan/product keys in the code and added notes to help future developers easily flip between live use and testing use. same thing for the Stripe api keys for production and testing too