:bug: finished training series CRUD abilities a 3 bugs
found the issue where the delete functionality for training series was broken. Found three other bugs. two where from code that didn't seem to have any use/function so after reveiwing the code with Adam we commented it out and left a message about in the file. the other bug was a quick fix in the trainingseriesPost file about a missing .data between props and id.