:bug: fixed the links on the landing, price, tams page
none of the links in the header/footer where working nor going to the right place. The issue appeared to come from the fact that Link was being used from React Router but in styled components it was an anchor tag and thus needed to be an href instead of a to = for the link to actually work. Also could turn the footer and nav into components since it's a lot of repeating code. Something to do in week 5 perhaps.