:bug: fixed links to github and linkedin for team page
when you click on a link to github, linkedIn, or portfolios it would open in a new tab but if you clikced on more than one, it would override that new tab. made a minor chang to allow multiple new tabs to be opened and added a line to each tag for security purposes as well.