:bug: Functionality restored for account selection
Account page functionality of displaying the correct info, being able to switch to a new subscription, unsubscribe (revert to "free") all restored and verified. userReducer code cleaned up quite a bit as well. For user switching their account type to be permanent, will need to open BE PR to update users table accordingly depending on which stripe action is taken.